Improve filter_labels method
Using ImgLabling.fromImageAndLabelSets() with a modified
LabelSets (i.e. removing entries we don't want) works but for
reasons I don't understand produces an ImgLabeling object that
has mis-matching indicies for the LabelSets and the size of the
array. In my testing, the size of the array is larger than the number
of elements in it (by 2). Perhaps trimToSize() isn't being called in
the right place? This object can't be used with an iterator due to
mismatching indicies. Thus I thought it might be better to not mess
with the ImgLabeling and instead convert the IndexImg into a new
template binary image for another CCA run.